sellQuantity

+++ src/main/java/com/spring/jwt/Interfaces/IProducts.java
@@ -24,9 +24,9 @@ public interface IProducts {
 
     List<ProductsDTO> getProductsByClothingType(String clothingType);
 
-    Integer getTotalStockQuantity();
+    Double getTotalStockQuantity();
 
-    Integer getStockQuantityByProductId(UUID productId);
+    Double getStockQuantityByProductId(UUID productId);
 
     List<ProductsDTO> saveProduct(List<ProductsDTO> productsDTOList);
 }
