+++ src/main/java/com/spring/jwt/dto/ProductsDTO.java
@@ -17,7 +17,7 @@ public class ProductsDTO {
 
     private ClothingType clothingType;
 
-    private List<Integer>stockQuantities;
+    private List<Double>stockQuantities;
 
 }
